#1183 refactoring
