+++ src/frontend/src/pages/TeamsPage/DeleteTeamModalContainer/DeleteTeamView.tsx
