+++ src/frontend/src/pages/TeamsPage/DescriptionPageBlock.tsx
